requirements.txt
accelerate==1.10.1
aenum==3.1.15
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiohttp-cors==0.8.1
aioice==0.10.1
aiortc==1.11.0
aiosignal==1.3.2
alembic==1.15.1
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
asgiref==3.8.1
astor==0.8.1
astroid==3.1.0
attrs==25.3.0
autopep8==2.1.0
av==14.3.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.13.3
black==24.4.2
blake3==1.0.5
boto3==1.37.20
botocore==1.37.20
bs4==0.0.2
build==1.2.2.post1
CacheControl==0.14.2
cachetools==5.5.2
cbor2==5.7.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
cleo==2.1.0
click==8.1.8
cloudpickle==3.1.1
clr_loader==0.2.7.post0
cobble==0.1.4
colorama==0.4.6
coloredlogs==15.0.1
compressed-tensors==0.10.2
crashtest==0.4.1
cryptography==44.0.2
dataclasses-json==0.6.7
deepgram-sdk==3.10.1
defusedxml==0.7.1
Deprecated==1.2.15
deprecation==2.1.0
depyf==0.19.0
dill==0.4.0
dirtyjson==1.0.8
diskcache==5.6.3
distlib==0.3.9
distro==1.9.0
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
dnspython==2.7.0
docling==2.51.0
docling-core==2.48.0
docling-ibm-models==3.9.1
docling-parse==4.4.0
dulwich==0.22.8
easyocr==1.7.2
ecdsa==0.19.0
einops==0.8.1
email_validator==2.2.0
emoji==2.14.1
et_xmlfile==2.0.0
eval_type_backport==0.2.2
Faker==37.6.0
fastapi==0.115.12
fastapi-cli==0.0.11
fastapi-cloud-cli==0.1.5
fastjsonschema==2.21.1
filelock==3.18.0
filetype==1.2.0
findpython==0.6.3
fire==0.7.0
firecrawl-py==1.12.0
flake8==7.0.0
flatbuffers==25.2.10
fonttools==4.56.0
frozendict==2.4.4
frozenlist==1.5.0
fsspec==2024.12.0
future==1.0.0
gguf==0.17.1
google-ai-generativelanguage==0.6.10
google-api-core==2.24.1
google-api-python-client==2.162.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-crc32c==1.7.1
google-generativeai==0.8.3
googleapis-common-protos==1.68.0
googlenewsdecoder==0.1.6
graphene==3.3
graphene-django==3.2.1
graphql-core==3.2.3
graphql-relay==3.2.0
greenlet==3.1.1
groq==0.20.0
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.34.4
humanfriendly==10.0
idna==3.10
ifaddr==0.2.0
imageio==2.37.0
installer==0.7.0
interegular==0.3.3
isort==5.13.2
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
joblib==1.4.2
jsonlines==3.1.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
keyring==25.6.0
langchain==0.3.21
langchain-community==0.3.20
langchain-core==0.3.48
langchain-experimental==0.0.58
langchain-openai==0.3.10
langchain-postgres==0.0.13
langchain-text-splitters==0.3.7
langdetect==1.0.9
langsmith==0.3.18
lark==1.2.2
latex2mathml==3.78.1
lazy_loader==0.4
linkify-it-py==2.0.3
llama-cloud==0.1.6
llama-index==0.12.11
llama-index-agent-openai==0.4.2
llama-index-cli==0.4.0
llama-index-core==0.12.11
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.3
llama-index-llms-openai==0.3.13
llama-index-multi-modal-llms-openai==0.4.2
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.3
llama-index-readers-llama-parse==0.4.0
llama-index-vector-stores-pinecone==0.4.2
llama-parse==0.5.19
llvmlite==0.44.0
lm-format-enforcer==0.10.12
loguru==0.7.3
lxml==5.3.1
Mako==1.3.9
Markdown==3.8
markdown-it-py==3.0.0
markdownify==0.14.1
marko==2.2.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mccabe==0.7.0
mdit-py-plugins==0.4.2
mdurl==0.1.2
mistral_common==1.8.4
more-itertools==10.6.0
mpire==2.10.2
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.19.0
multidict==6.2.0
multiprocess==0.70.18
multitasking==0.0.11
mypy==1.10.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.13.0
nltk==3.9.1
numba==0.61.2
numpy==2.2.4
odfpy==1.4.1
olefile==0.47
onnxruntime==1.20.1
openai==1.107.0
openai-harmony==0.0.4
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
orjson==3.10.16
outcome==1.3.0.post0
outlines_core==0.2.10
packaging==24.2
pandas==2.2.3
partial-json-parser==0.2.1.1.post6
passlib==1.7.4
pathspec==0.12.1
pbs-installer==2025.3.17
pdf2docx==0.5.8
pdf2image==1.17.0
peewee==3.17.5
pgvector==0.3.6
pillow==11.1.0
pinecone==5.4.2
pinecone-client==4.1.2
pinecone-plugin-inference==3.1.0
pinecone-plugin-interface==0.0.7
pipecat-ai==0.0.63
pipecat-ai-small-webrtc-prebuilt==0.5.0
pkginfo==1.12.1.2
platformdirs==4.2.2
pluggy==1.6.0
plum-dispatch==1.7.4
poetry==2.1.2
poetry-core==2.1.2
polyfactory==2.22.2
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.22.1
promise==2.3
propcache==0.3.1
proto-plus==1.26.0
protobuf==5.29.3
psutil==7.0.0
psycopg==3.2.6
psycopg-binary==3.2.6
psycopg-pool==3.2.6
psycopg2==2.9.10
psycopg2-binary==2.9.9
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pybase64==1.4.2
pyclipper==1.3.0.post6
pycodestyle==2.11.1
pycountry==24.6.1
pycparser==2.22
pydantic==2.11.7
pydantic-extra-types==2.10.5
pydantic-settings==2.8.1
pydantic_core==2.33.2
pyee==13.0.0
pyflakes==3.2.0
Pygments==2.19.1
pylatexenc==2.10
pylibsrtp==0.12.0
pylint==3.1.0
pyloudnorm==0.1.1
PyMuPDF==1.26.4
pymupdf4llm==0.0.27
pyOpenSSL==25.0.0
pyparsing==3.2.1
pypdf==5.4.0
PyPDF2==3.0.1
pypdfium2==4.30.0
pyproject_hooks==1.2.0
pyreadline3==3.5.4
PySocks==1.7.1
pytesseract==0.3.13
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-iso639==2025.2.18
python-jose==3.3.0
python-json-logger==3.3.0
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
python-pptx==1.0.2
pythonnet==3.0.5
pytz==2025.2
pywin32==311
pywin32-ctypes==0.2.3
PyYAML==6.0.2
pyzmq==27.1.0
RapidFuzz==3.12.2
redis==5.0.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
resampy==0.4.3
rich==13.9.4
rich-toolkit==0.15.1
rignore==0.6.4
rpds-py==0.27.1
rsa==4.9
rtree==1.4.1
s3transfer==0.11.4
safetensors==0.6.2
scikit-image==0.25.2
scipy==1.15.2
selectolax==0.3.27
selenium==4.20.0
semchunk==2.2.2
sentencepiece==0.2.1
sentry-sdk==2.37.1
setproctitle==1.3.7
setuptools==79.0.1
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
soupsieve==2.6
soxr==0.5.0.post1
spire-doc==13.3.7
SQLAlchemy==2.0.39
sqlparse==0.5.0
srt==3.5.3
starlette==0.46.1
striprtf==0.0.29
structlog==24.1.0
sympy==1.13.3
tabulate==0.9.0
tenacity==9.0.0
termcolor==2.5.0
tesseract==0.1.3
text-unidecode==1.3
textual==1.0.0
tifffile==2025.9.9
tiktoken==0.9.0
tokenizers==0.22.0
tomlkit==0.12.5
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
transformers==4.56.1
trio==0.25.1
trio-websocket==0.11.1
trove-classifiers==2025.3.19.19
typer==0.16.1
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
tzlocal==5.3.1
uc-micro-py==1.0.3
ujson==5.10.0
unstructured==0.17.2
unstructured-client==0.31.5
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
virtualenv==20.30.0
vllm==0.10.1.1
vosk==0.3.45
watchfiles==1.0.5
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
win32_setctime==1.2.0
wrapt==1.17.2
wsproto==1.2.0
xlrd==2.0.1
XlsxWriter==3.2.2
yarl==1.18.3
yfinance==0.2.40
zstandard==0.23.0
